requirements.txt
pillow==7.1.2